Updated module augmentation for html-embed package.
