+++ packages/ckeditor5-html-embed/src/htmlembedui.ts
@@ -61,9 +61,3 @@ export default class HtmlEmbedUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PLuginsMap {
-		[ HtmlEmbedUI.pluginName ]: HtmlEmbedUI;
-	}
-}
